# CXX-3094 Upgrade Catch2 from v2 to v3