--- a/THIRD-PARTY-NOTICES
+++ b/THIRD-PARTY-NOTICES
@@ -15,7 +15,7 @@ https://github.com/mongodb/mongo-cxx-driver.
 ---------------------------------------
 [Note: Catch is used only for testing.]
 
-Boost Software License - Version 1.0 - August 17th, 2003
+Copyright Catch2 Authors
 
 Distributed under the Boost Software License, Version 1.0.
 See http://www.boost.org/LICENSE_1_0.txt