--- a/src/bsoncxx/test/bson_types.cpp
+++ b/src/bsoncxx/test/bson_types.cpp
@@ -18,9 +18,10 @@
 #include <bsoncxx/builder/basic/document.hpp>
 #include <bsoncxx/document/element.hpp>
 #include <bsoncxx/stdx/string_view.hpp>
-#include <bsoncxx/test/catch.hh>
 #include <bsoncxx/types.hpp>
 #include <bsoncxx/types/bson_value/view.hpp>
+
+#include <bsoncxx/test/catch.hh>
 
 namespace {