--- a/src/bsoncxx/test/bson_types.cpp
+++ b/src/bsoncxx/test/bson_types.cpp
@@ -23,6 +23,8 @@
 
 #include <bsoncxx/test/catch.hh>
 
+#include <catch2/matchers/catch_matchers_string.hpp>
+
 namespace {
 
 using namespace bsoncxx;
@@ -154,7 +156,7 @@ TEST_CASE("b_int32", "[bsoncxx::type::b_int32]") {
     REQUIRE(!(a == c));
 }
 
-TEST_CASE("b_timestamp", "[bsoncxx::type::b_timestamp") {
+TEST_CASE("b_timestamp", "[bsoncxx::type::b_timestamp]") {
     b_timestamp a{100, 1000};
     b_timestamp b{100, 1000};
     b_timestamp c{200, 1000};
@@ -374,22 +376,25 @@ TEST_CASE("document uninitialized element throws exceptions", "") {
     bsoncxx::document::value doc = make_document(kvp("foo", "bar"));
 
     REQUIRE_THROWS_WITH(doc["doesnotexist"].get_string().value,
-                        Catch::Contains("cannot get string from an uninitialized element with key "
-                                        "\"doesnotexist\": unset document::element"));
+
+                        Catch::Matchers::ContainsSubstring(
+                            "cannot get string from an uninitialized element with key "
+                            "\"doesnotexist\": unset document::element"));
 
     REQUIRE_THROWS_WITH(doc["alsodoesnotexist"].get_value(),
-                        Catch::Contains("cannot return the type of uninitialized element with key "
-                                        "\"alsodoesnotexist\": unset document::element"));
+                        Catch::Matchers::ContainsSubstring(
+                            "cannot return the type of uninitialized element with key "
+                            "\"alsodoesnotexist\": unset document::element"));
 
     // Ensure a non-existing element evaluates to false.
     REQUIRE(!doc["doesnotexist"]);
     // Ensure finding a non-existing element results in an end iterator.
     REQUIRE(doc.find("doesnotexist") == doc.cend());
     // Ensure getting a key from a non-existing element results in an exception.
-    REQUIRE_THROWS_WITH(
-        doc["doesnotexist"].key(),
-        Catch::Contains("cannot return the key from an uninitialized element with key "
-                        "\"doesnotexist\": unset document::element"));
+    REQUIRE_THROWS_WITH(doc["doesnotexist"].key(),
+                        Catch::Matchers::ContainsSubstring(
+                            "cannot return the key from an uninitialized element with key "
+                            "\"doesnotexist\": unset document::element"));
 }
 
 TEST_CASE("array uninitialized element throws exceptions", "") {
@@ -398,16 +403,17 @@ TEST_CASE("array uninitialized element throws exceptions", "") {
     bsoncxx::array::value arr = make_array("a", "b", "c");
 
     REQUIRE_THROWS_WITH(arr.view()[3].get_string().value,
-                        Catch::Contains("cannot get string from an uninitialized element with key "
-                                        "\"3\": unset document::element"));
+                        Catch::Matchers::ContainsSubstring(
+                            "cannot get string from an uninitialized element with key "
+                            "\"3\": unset document::element"));
     // Ensure a non-existing element evaluates to false.
     REQUIRE(!arr.view()[3]);
     // Ensure finding a non-existing element results in an end iterator.
     REQUIRE(arr.view().find(3) == arr.view().cend());
     // Ensure getting a key from a non-existing element results in an exception.
-    REQUIRE_THROWS_WITH(
-        arr.view()[3].key(),
-        Catch::Contains("cannot return the key from an uninitialized element with key "
-                        "\"3\": unset document::element"));
+    REQUIRE_THROWS_WITH(arr.view()[3].key(),
+                        Catch::Matchers::ContainsSubstring(
+                            "cannot return the key from an uninitialized element with key "
+                            "\"3\": unset document::element"));
 }
 }  // namespace