--- a/src/bsoncxx/test/bson_util_itoa.cpp
+++ b/src/bsoncxx/test/bson_util_itoa.cpp
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #include <bsoncxx/private/itoa.hh>
+
 #include <bsoncxx/test/catch.hh>
 
 namespace {