--- a/src/bsoncxx/test/oid.cpp
+++ b/src/bsoncxx/test/oid.cpp
@@ -21,6 +21,7 @@
 
 #include <bsoncxx/oid.hpp>
 #include <bsoncxx/private/libbson.hh>
+
 #include <bsoncxx/test/catch.hh>
 
 using namespace bsoncxx;