--- a/src/bsoncxx/test/view_or_value.cpp
+++ b/src/bsoncxx/test/view_or_value.cpp
@@ -19,6 +19,7 @@
 #include <bsoncxx/document/view_or_value.hpp>
 #include <bsoncxx/json.hpp>
 #include <bsoncxx/string/view_or_value.hpp>
+
 #include <bsoncxx/test/catch.hh>
 
 namespace {