--- a/src/mongocxx/test/change_streams.cpp
+++ b/src/mongocxx/test/change_streams.cpp
@@ -18,7 +18,6 @@
 #include <bsoncxx/builder/basic/document.hpp>
 #include <bsoncxx/document/value.hpp>
 #include <bsoncxx/private/libbson.hh>
-#include <bsoncxx/test/catch.hh>
 #include <mongocxx/client.hpp>
 #include <mongocxx/collection.hpp>
 #include <mongocxx/instance.hpp>
@@ -27,8 +26,11 @@
 #include <mongocxx/pipeline.hpp>
 #include <mongocxx/pool.hpp>
 #include <mongocxx/private/libbson.hh>
-#include <mongocxx/test/client_helpers.hh>
 #include <mongocxx/write_concern.hpp>
+
+#include <bsoncxx/test/catch.hh>
+#include <mongocxx/test/client_helpers.hh>
+
 #include <third_party/catch/include/helpers.hpp>
 
 namespace {