--- a/src/mongocxx/test/client.cpp
+++ b/src/mongocxx/test/client.cpp
@@ -13,18 +13,20 @@
 // limitations under the License.
 
 #include <bsoncxx/string/to_string.hpp>
-#include <bsoncxx/test/catch.hh>
 #include <mongocxx/client.hpp>
 #include <mongocxx/exception/logic_error.hpp>
 #include <mongocxx/instance.hpp>
 #include <mongocxx/pool.hpp>
 #include <mongocxx/private/conversions.hh>
 #include <mongocxx/private/libmongoc.hh>
-#include <mongocxx/test/client_helpers.hh>
 #include <mongocxx/uri.hpp>
-#include <third_party/catch/include/helpers.hpp>
 
 #include <mongocxx/config/private/prelude.hh>
+
+#include <bsoncxx/test/catch.hh>
+#include <mongocxx/test/client_helpers.hh>
+
+#include <third_party/catch/include/helpers.hpp>
 
 namespace {
 using namespace mongocxx;