--- a/src/mongocxx/test/collection_mocked.cpp
+++ b/src/mongocxx/test/collection_mocked.cpp
@@ -35,8 +35,7 @@
 #include <mongocxx/read_preference.hpp>
 
 #include <bsoncxx/test/catch.hh>
-
-#include <third_party/catch/include/helpers.hpp>
+#include <mongocxx/test/catch_helpers.hh>
 
 namespace {
 using namespace mongocxx;