… .mci.yml
… CHANGELOG.md
… cmake/CMakeLists.txt
… src/CMakeLists.txt
… src/bsoncxx/test/CMakeLists.txt
… src/mongocxx/test/CMakeLists.txt
… src/mongocxx/test/collection.cpp
… src/mongocxx/test/conversions.cpp
… src/mongocxx/test/database.cpp
… src/mongocxx/test/gridfs/bucket.cpp
… src/mongocxx/test/gridfs/downloader.cpp
… src/mongocxx/test/gridfs/uploader.cpp
… src/mongocxx/test/hint.cpp
… src/mongocxx/test/index_view.cpp
… src/mongocxx/test/instance.cpp
… src/mongocxx/test/logging.cpp
… src/mongocxx/test/model/delete_many.cpp
… src/mongocxx/test/model/delete_one.cpp
… src/mongocxx/test/model/insert_one.cpp
… src/mongocxx/test/model/replace_one.cpp
… src/mongocxx/test/model/update_many.cpp
… src/mongocxx/test/model/update_one.cpp
… src/mongocxx/test/options/aggregate.cpp
… src/mongocxx/test/options/bulk_write.cpp
… src/mongocxx/test/options/client_session.cpp
… src/mongocxx/test/options/count.cpp
… src/mongocxx/test/options/create_collection.cpp
… src/mongocxx/test/options/delete.cpp
… src/mongocxx/test/options/distinct.cpp
… src/mongocxx/test/options/find.cpp