@eramongodb 2024-08-28
Although not directly related, given the Catch2 migration improves trailing semicolon handling + relocation/modification of the newly-named `catch_helpers.hh` header, applied a drive-by Reduce Warnings commit that addresses 294 (174 unique) `-Wextra-semi` and `-Wextra-semi-stmt` warnings with Clang.